+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
   Sunqua Peak CM fight.
 - `Log::gadgets` to retrieve all gadget agents.
 - `Log::build_id` to retrieve the game's build id.
+- The `serde` optional feature to enable (de)serialization of API types.
 
 ### Changed
 - `gamedata::Boss` has been split in `gamedata::Boss` and `gamedata::Encounter`